[ refactor ] switched form elements to spans.
This is a stateless application. All the configuration data is derived from the Book.toml file. Presenting that information as if it was a form makes the user think they can change it there. This breaks the idea that the files on disk are the source of truth and could lead to "out of sync" problems between whatever is on the form and the data.