[[ site ]] add licensing property
A new `metadata.license` property has been added which is a string and can be HTML. This can be used to add license information to the Website, such as adding Creative Commons logo and link. This doesn't affect the EPUB. The author needs to add their copyright page as it's own file in the frontmatter.