lots of bugs
Internal links were broken because epubs expect XHTMLs and the web-version wanted HTMLs. It took me a while to solve this.

There is a new property of `site` config that allows adding the name of a PDF file for download.

```
[site]
downloadPDF = my-book.pdf
```

Be aware that LWP can't generate PDFs, you need to generate them on your own and place them with the correct name inside the `files/` folder.